+++ cvmfs/libcvmfs.cc
@@ -28,275 +28,6 @@
 
 using namespace std;  // NOLINT
 
-int set_option(char const *name, char const *value, bool *var) {
-  if (*value != '\0') {
-    fprintf(stderr, "Option %s=%s contains a value when none was expected.\n",
-            name, value);
-    return -1;
-  }
-  *var = true;
-  return 0;
-}
-
-int set_option(char const *name, char const *value, unsigned *var) {
-  unsigned v = 0;
-  int end = 0;
-  int rc = sscanf(value, "%u%n", &v, &end);
-  if (rc != 1 || value[end] != '\0') {
-    fprintf(stderr, "Invalid unsigned integer value for %s=%s\n", name, value);
-    return -1;
-  }
-  *var = v;
-  return 0;
-}
-
-int set_option(char const *name, char const *value, uint64_t *var) {
-  uint64_t v = 0;
-  int end = 0;
-  int rc = sscanf(value, "%" PRIu64 "%n", &v, &end);
-  if (rc != 1 || value[end] != '\0') {
-    fprintf(stderr, "Invalid unsigned long integer value for %s=%s\n",
-            name, value);
-    return -1;
-  }
-  *var = v;
-  return 0;
-}
-
-int set_option(char const *name, char const *value, int *var) {
-  int v = 0;
-  int end = 0;
-  int rc = sscanf(value, "%d%n", &v, &end);
-  if (rc != 1 || value[end] != '\0') {
-    fprintf(stderr, "Invalid integer value for %s=%s\n", name, value);
-    return -1;
-  }
-  *var = v;
-  return 0;
-}
-
-int set_option(char const *name, char const *value, string *var) {
-  *var = value;
-  return 0;
-}
-
-
-#define CVMFS_OPT(var) if (strcmp(name, #var) == 0) \
-  return ::set_option(name, value, &var)
-
-struct cvmfs_repo_options : public LibContext::options {
-  int set_option(char const *name, char const *value) {
-    CVMFS_OPT(allow_unsigned);
-    CVMFS_OPT(blacklist);
-    CVMFS_OPT(deep_mount);  // deprecated
-    CVMFS_OPT(fallback_proxies);
-    CVMFS_OPT(mountpoint);
-    CVMFS_OPT(proxies);
-    CVMFS_OPT(pubkey);
-    CVMFS_OPT(repo_name);
-    CVMFS_OPT(timeout);
-    CVMFS_OPT(timeout_direct);
-    CVMFS_OPT(tracefile);
-    CVMFS_OPT(url);
-
-    fprintf(stderr, "Unknown repo option: %s\n", name);
-    return -1;
-  }
-
-  int verify_sanity() {
-    if (mountpoint.empty() && !repo_name.empty()) {
-      mountpoint = "/cvmfs/";
-      mountpoint += repo_name;
-    }
-    while (mountpoint.length() > 0 && mountpoint[mountpoint.length()-1] == '/')
-    {
-      mountpoint.resize(mountpoint.length()-1);
-    }
-
-    return LIBCVMFS_FAIL_OK;
-  }
-};
-
-struct cvmfs_global_options : public LibGlobals::options {
-  int set_option(char const *name, char const *value) {
-    CVMFS_OPT(alien_cache);
-    CVMFS_OPT(alien_cachedir);
-    CVMFS_OPT(cache_directory);
-    CVMFS_OPT(cachedir);
-    CVMFS_OPT(lock_directory);
-    CVMFS_OPT(change_to_cache_directory);
-    CVMFS_OPT(logfile);
-    CVMFS_OPT(log_file);
-    CVMFS_OPT(log_prefix);
-    CVMFS_OPT(log_syslog_level);
-    CVMFS_OPT(syslog_level);
-    CVMFS_OPT(max_open_files);
-    CVMFS_OPT(nofiles);
-    CVMFS_OPT(quota_limit);
-    CVMFS_OPT(quota_threshold);
-    CVMFS_OPT(rebuild_cachedb);
-
-    fprintf(stderr, "Unknown global option: %s\n", name);
-    return LIBCVMFS_FAIL_BADOPT;
-  }
-
-  int verify_sanity() {
-    // Alias handling
-    if ((nofiles >= 0) && (max_open_files != 0) && (nofiles != max_open_files))
-      return LIBCVMFS_FAIL_BADOPT;
-    if (nofiles >= 0)
-      max_open_files = nofiles;
-
-    if ((syslog_level >= 0) && (log_syslog_level != 0) &&
-        (syslog_level != log_syslog_level))
-    {
-      return LIBCVMFS_FAIL_BADOPT;
-    }
-    if (syslog_level >= 0)
-      log_syslog_level = syslog_level;
-    if (log_syslog_level < 0)
-      log_syslog_level = 3;
-
-    if ((logfile != "") && (log_file != "") && (log_file != logfile))
-      return LIBCVMFS_FAIL_BADOPT;
-    if (logfile != "")
-      log_file = logfile;
-
-    if ((cachedir != "") && (cache_directory != "") &&
-        (cache_directory != cachedir))
-    {
-      return LIBCVMFS_FAIL_BADOPT;
-    }
-    if (cachedir != "")
-      cache_directory = cachedir;
-
-    return LIBCVMFS_FAIL_OK;
-  }
-};
-
-
-/**
- * Structure to parse the file system options.
- */
-template <class DerivedT>
-struct cvmfs_options : public DerivedT {
-  int set_option(char const *name, char const *value) {
-    return DerivedT::set_option(name, value);
-  }
-
-  int parse_options(char const *options)
-  {
-    while (*options) {
-      char const *next = options;
-      string name;
-      string value;
-
-      // get the option name
-      for (next=options; *next && *next != ',' && *next != '='; next++) {
-        if (*next == '\\') {
-          next++;
-          if (*next == '\0') break;
-        }
-        name += *next;
-      }
-
-      if (*next == '=') {
-        next++;
-      }
-
-      // get the option value
-      for (; *next && *next != ','; next++) {
-        if (*next == '\\') {
-          next++;
-          if (*next == '\0') break;
-        }
-        value += *next;
-      }
-
-      if (!name.empty() || !value.empty()) {
-        int result = set_option(name.c_str(), value.c_str());
-        if (result != 0) {
-          return result;
-        }
-      }
-
-      if (*next == ',') next++;
-      options = next;
-    }
-
-    return DerivedT::verify_sanity();
-  }
-};
-
-typedef cvmfs_options<cvmfs_repo_options>   repo_options;
-typedef cvmfs_options<cvmfs_global_options> global_options;
-
-/**
- * Display the usage message.
- */
-static void usage() {
-  struct cvmfs_repo_options defaults;
-  fprintf(stderr,
-  "CernVM-FS version %s\n"
-  "Copyright (c) 2009- CERN\n"
-  "All rights reserved\n\n"
-  "Please visit http://cernvm.cern.ch/project/info for license details "
-  "and author list.\n\n"
-
-  "libcvmfs options are expected in the form: option1,option2,option3,...\n"
-  "Within an option, the characters , and \\ must be preceded by \\.\n\n"
-
-  "There are two types of options (global and repository specifics)\n"
-  "  cvmfs_init()        expects global options\n"
-  "  cvmfs_attach_repo() expects repository specific options\n"
-
-  "global options are:\n"
-  " cache_directory/cachedir=DIR Where to store disk cache\n"
-  " change_to_cache_directory  Performs a cd to the cache directory "
-                               "(performance tweak)\n"
-  " alien_cache                Treat cache directory as alien cache\n"
-  " alien_cachedir=DIR         Explicitly set an alien cache directory\n"
-  " lock_directory=DIR         Directory for per instance lock files.\n"
-  "                            Needs to be on a file system with POSIX locks.\n"
-  "                            Should be different from alien cache directory."
-  "                            \nDefaults to cache_directory.\n"
-  " (log_)syslog_level=LEVEL   Sets the level used for syslog to "
-                               "DEBUG (1), INFO (2), or NOTICE (3).\n"
-  "                            Default is NOTICE.\n"
-  " log_prefix                 String to use as a log prefix in syslog\n"
-  " log_file/logfile           Logs all messages to FILE instead of "
-                               "stderr and daemonizes.\n"
-  "                            Makes only sense for the debug version\n"
-  " nofiles/max_open_files     Set the maximum number of open files "
-                               "for CernVM-FS process (soft limit)\n\n"
-
-  "repository specific options are:"
-  " repo_name=REPO_NAME        Unique name of the mounted repository, "
-                               "e.g. atlas.cern.ch\n"
-  " url=REPOSITORY_URL         The URL of the CernVM-FS server(s): "
-                               "'url1;url2;...'\n"
-  " timeout=SECONDS            Timeout for network operations (default is %d)\n"
-  " timeout_direct=SECONDS     Timeout for network operations without proxy "
-                               "(default is %d)\n"
-  " proxies=HTTP_PROXIES       Set the HTTP proxy list, such as "
-                               "'proxy1|proxy2;DIRECT'\n"
-  " fallback_proxies=PROXIES   Set the fallback proxy list, such as "
-                               "'proxy1;proxy2'\n"
-  " tracefile=FILE             Trace FUSE opaerations into FILE\n"
-  " pubkey=PEMFILE             Public RSA key that is used to verify the "
-                               "whitelist signature.\n"
-  " allow_unsigned             Accept unsigned catalogs "
-                               "(allows man-in-the-middle attacks)\n"
-  " deep_mount=PREFIX          Path prefix if a repository is mounted on a "
-                               "nested catalog,\n"
-  "                            i.e. deep_mount=/software/15.0.1\n"
-  " mountpoint=PATH            Path to root of repository, "
-                               "e.g. /cvmfs/atlas.cern.ch\n"
-  " blacklist=FILE             Local blacklist for invalid certificates. "
-                               "Has precedence over the whitelist.\n",
-  PACKAGE_VERSION, defaults.timeout, defaults.timeout_direct);
-}
-
 /**
  * Expand symlinks in all levels of a path.  Also, expand ".." and ".".  This
  * also has the side-effect of ensuring that cvmfs_getattr() is called on all
@@ -385,8 +116,8 @@ static int expand_path(
   }
   if (ln_buf[0] == '/') {
     // symlink is absolute path, strip /cvmfs/$repo
-    unsigned len = ctx->mountpoint().length();
-    if (strncmp(ln_buf, ctx->mountpoint().c_str(), len) == 0 &&
+    unsigned len = ctx->mount_point()->fqrn().length();
+    if (strncmp(ln_buf, ctx->mount_point()->fqrn().c_str(), len) == 0 &&
         (ln_buf[len] == '/' || ln_buf[len] == '\0'))
     {
       buf = ln_buf+len;
@@ -397,7 +128,7 @@ static int expand_path(
       LogCvmfs(kLogCvmfs, kLogDebug,
                "libcvmfs can't resolve symlinks to paths outside of the repo: "
                "%s --> %s (mountpoint=%s)",
-               path, ln_buf, ctx->mountpoint().c_str());
+               path, ln_buf, ctx->mount_point()->fqrn().c_str());
       errno = ENOENT;
       return -1;
     }
@@ -568,29 +299,6 @@ int cvmfs_listdir(
 }
 
 
-LibContext* cvmfs_attach_repo(char const *options)
-{
-  // Parse options
-  repo_options opts;
-  int parse_result = opts.parse_options(options);
-  if (parse_result != 0)
-  {
-    if (parse_result < 0) {
-      fprintf(stderr, "Invalid CVMFS options: %s.\n", options);
-      usage();
-    }
-    return NULL;
-  }
-  if (opts.url.empty()) {
-    fprintf(stderr, "No url specified in CVMFS repository options: %s.\n",
-            options);
-    return NULL;
-  }
-
-  return LibContext::Create(opts);
-}
-
-
 cvmfs_errors cvmfs_attach_repo_v2(
   const char *fqrn,
   OptionsManager *opts,
@@ -604,20 +312,7 @@ cvmfs_errors cvmfs_attach_repo_v2(
 
 
 void cvmfs_detach_repo(LibContext *ctx) {
-  LibContext::Destroy(ctx);
-}
-
-
-int cvmfs_init(char const *options) {
-  global_options opts;
-  int parse_result = opts.parse_options(options);
-  if (parse_result != 0) {
-    fprintf(stderr, "Invalid CVMFS global options: %s.\n", options);
-    usage();
-    return parse_result;
-  }
-
-  return LibGlobals::Initialize(opts);
+  delete ctx;
 }
